Move GCSettings to shared
Mainly just type definitions and error handling that gets shared.
